git-svn-id: svn://localhost/public/freewpc/trunk@32 5087d79f-0ef8-0310-a22a-8fd82e81f20f

+++ include/sys/sound.h
@@ -12,7 +12,7 @@
 #define SND_SCROLL				0x59
 
 #ifndef __SASM__
-typedef uint8_t music_code_t, sound_code_t;
+typedef uint16_t music_code_t, sound_code_t;
 
 void music_off (void);
 void music_set (music_code_t code);
